refactor(web): complete WorkspaceTopBar component/hook extraction (phase 2)
Split the remaining ~1150-line WorkspaceTopBar.tsx into focused,
under-400-line modules under workspace-top-bar/: CanvasItem,
CanvasDropdown, CanvasActionsMenu, VersionPanel, NewCanvasDialog, and
TopBarSecondaryActions for presentation, plus useCanvasNames (single
owner of the workspaceNamesSchema-derived state and its writers),
useSaveVersion, useCopyCanvasUrl, useSceneExport, useCanvasRename, and
useCreateCanvas for behavior. WorkspaceTopBar.tsx is now a ~390-line
orchestrator with the same default export, Props shape, and
WorkspaceTopBarCapabilities contract.

Adds a browser regression covering the aria-hidden-after-close bug
this file previously fixed for the Canvas actions menu, plus
nearest-layer tests for the newly extracted hooks (schema-parse
handling, unmount-during-pending-op guards, and concurrent-save
dedup). Existing WorkspaceTopBar.test.tsx (46 tests) and
WorkspaceTopBar.browser.test.tsx (11 pre-existing + 1 new) assertions
are preserved unweakened.
